refactor(tests): Move all tests out of mod
This is prep for merged complete/streaming.

As I expect all `streaming` and `complete` mods to be deprecated, I've
isolated there tests to those mods.  When we add the merged versions,
that'll be too many tests for the `mod.rs` files (and requires more
recompilation), so I'm moving these out to tests.rs.
